+++ studio/src/components/ui/spacer.tsx
@@ -1,3 +1,5 @@
-export const Spacer = () => {
-  return <div className="flex-1" />;
+import { cn } from "@/lib/utils";
+
+export const Spacer = ({ className }: { className?: string }) => {
+  return <div className={cn("flex-1", className)} />;
 };
